Moving Join Meeting API from calendar to meeting
